added superbase class called generic
