+++ lib/evertils/common/query.rb
@@ -1,15 +1,7 @@
 module Evertils
   module Common
     module Query
-      class Base
-
-        #
-        # @since 0.2.8
-        def deprecation_notice(version, message)
-          output = "Deprecated as of #{version}"
-          output += "\nReason: #{message}" if message
-          output
-        end
+      class Base < Common::Generic
 
         #
         # @since 0.2.8
